重写修改照片的逻辑

+++ controllers/userManager.js
@@ -82,7 +82,13 @@ exports.modify = function(req,res){
                 	if (!fs.existsSync(savePath)) {
                     	fs.mkdirSync(savePath);
                 	}
-                		fs.rename(uploadedPath, savePath + 'headPic.jpg');
+                	fs.exists(savePath + 'headPic.jpg', function( exists ){
+    					if(exists){
+    						fs.writeFileSync(uploadedPath, fs.readFileSync(savePath + 'headPic.jpg'));
+    					}else{
+    						fs.rename(uploadedPath, savePath + 'headPic.jpg');
+    					}
+					});
             		}else{
                 		userinfo.picture = 'false';
             		}
